+++ src/impls.rs
@@ -1660,7 +1660,8 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                Self::NonAtomicType::from_bits(self.0.fetch_add(value.to_bits(), order))
+                self.fetch_update(Ordering::Relaxed, order, |x| Some(x + value))
+                    .unwrap()
             }
 
             #[inline(always)]
@@ -1669,7 +1670,8 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                Self::NonAtomicType::from_bits(self.0.fetch_sub(value.to_bits(), order))
+                self.fetch_update(Ordering::Relaxed, order, |x| Some(x - value))
+                    .unwrap()
             }
 
             #[inline(always)]
@@ -1678,7 +1680,10 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                Self::NonAtomicType::from_bits(self.0.fetch_min(value.to_bits(), order))
+                self.fetch_update(Ordering::Relaxed, order, |x| {
+                    Some(Self::NonAtomicType::min(x, value))
+                })
+                .unwrap()
             }
 
             #[inline(always)]
@@ -1687,7 +1692,10 @@ macro_rules! impl_f16 {
                 value: Self::NonAtomicType,
                 order: Ordering,
             ) -> Self::NonAtomicType {
-                Self::NonAtomicType::from_bits(self.0.fetch_max(value.to_bits(), order))
+                self.fetch_update(Ordering::Relaxed, order, |x| {
+                    Some(Self::NonAtomicType::max(x, value))
+                })
+                .unwrap()
             }
         }
 
@@ -1786,19 +1794,8 @@ macro_rules! impl_f16 {
                 set_order: Ordering,
                 fetch_order: Ordering,
             ) -> Self::NonAtomicType {
-                loop {
-                    let orig = self.load(fetch_order);
-                    let res = orig.saturating_add(value);
-                    if res == orig {
-                        break res;
-                    }
-                    if self
-                        .compare_exchange_weak(orig, res, set_order, fetch_order)
-                        .is_ok()
-                    {
-                        break res;
-                    }
-                }
+                self.fetch_update(set_order, fetch_order, |x| Some(x.saturating_add(value)))
+                    .unwrap()
             }
             #[inline(always)]
             fn fetch_saturating_sub(
@@ -1807,19 +1804,8 @@ macro_rules! impl_f16 {
                 set_order: Ordering,
                 fetch_order: Ordering,
             ) -> Self::NonAtomicType {
-                loop {
-                    let orig = self.load(fetch_order);
-                    let res = orig.saturating_sub(value);
-                    if res == orig {
-                        break res;
-                    }
-                    if self
-                        .compare_exchange_weak(orig, res, set_order, fetch_order)
-                        .is_ok()
-                    {
-                        break res;
-                    }
-                }
+                self.fetch_update(set_order, fetch_order, |x| Some(x.saturating_sub(value)))
+                    .unwrap()
             }
             #[inline(always)]
             fn fetch_saturating_mul(
@@ -1828,19 +1814,8 @@ macro_rules! impl_f16 {
                 set_order: Ordering,
                 fetch_order: Ordering,
             ) -> Self::NonAtomicType {
-                loop {
-                    let orig = self.load(fetch_order);
-                    let res = orig.saturating_mul(value);
-                    if res == orig {
-                        break res;
-                    }
-                    if self
-                        .compare_exchange_weak(orig, res, set_order, fetch_order)
-                        .is_ok()
-                    {
-                        break res;
-                    }
-                }
+                self.fetch_update(set_order, fetch_order, |x| Some(x.saturating_mul(value)))
+                    .unwrap()
             }
             #[inline(always)]
             fn fetch_saturating_div(
@@ -1849,19 +1824,8 @@ macro_rules! impl_f16 {
                 set_order: Ordering,
                 fetch_order: Ordering,
             ) -> Self::NonAtomicType {
-                loop {
-                    let orig = self.load(fetch_order);
-                    let res = orig.saturating_div(value);
-                    if res == orig {
-                        break res;
-                    }
-                    if self
-                        .compare_exchange_weak(orig, res, set_order, fetch_order)
-                        .is_ok()
-                    {
-                        break res;
-                    }
-                }
+                self.fetch_update(set_order, fetch_order, |x| Some(x.saturating_div(value)))
+                    .unwrap()
             }
             #[cfg(feature = "std")]
             #[inline(always)]
@@ -1871,19 +1835,8 @@ macro_rules! impl_f16 {
                 set_order: Ordering,
                 fetch_order: Ordering,
             ) -> Self::NonAtomicType {
-                loop {
-                    let orig = self.load(fetch_order);
-                    let res = orig.saturating_pow(value);
-                    if res == orig {
-                        break res;
-                    }
-                    if self
-                        .compare_exchange_weak(orig, res, set_order, fetch_order)
-                        .is_ok()
-                    {
-                        break res;
-                    }
-                }
+                self.fetch_update(set_order, fetch_order, |x| Some(x.saturating_pow(value)))
+                    .unwrap()
             }
         }